[#2291]:+ Fixed inconsistent plugin placement, Moved all the plugins to Plugin as root. Removed duplicate di plugin references.
