+++ app/code/Magento/CatalogUrlRewrite/Plugin/Model/Category/Move.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\CatalogUrlRewrite\Model\Category\Plugin\Category;
+namespace Magento\CatalogUrlRewrite\Plugin\Model\Category;
 
 use Magento\Catalog\Model\Category;
 use Magento\CatalogUrlRewrite\Model\CategoryUrlPathGenerator;
