+++ app/code/Magento/CatalogUrlRewrite/Plugin/Model/Category/Remove.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\CatalogUrlRewrite\Model\Category\Plugin\Category;
+namespace Magento\CatalogUrlRewrite\Plugin\Model\Category;
 
 use Magento\UrlRewrite\Service\V1\Data\UrlRewrite;
