+++ app/code/Magento/CatalogUrlRewrite/Plugin/Model/Category/Storage.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\CatalogUrlRewrite\Model\Category\Plugin;
+namespace Magento\CatalogUrlRewrite\Plugin\Model\Category;
 
 use Magento\CatalogUrlRewrite\Model\ProductUrlRewriteGenerator;
 use Magento\UrlRewrite\Model\StorageInterface;
