+++ app/code/Magento/CatalogUrlRewrite/etc/adminhtml/di.xml
@@ -7,14 +7,10 @@
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <type name="Magento\Store\Model\ResourceModel\Store">
-        <plugin name="store_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Store\View"/>
+        <plugin name="store_plugin" type="Magento\CatalogUrlRewrite\Plugin\Model\Category\Store\View"/>
     </type>
     <type name="Magento\Store\Model\ResourceModel\Group">
-        <plugin name="group_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Store\Group"/>
-    </type>
-    <type name="Magento\Catalog\Model\ResourceModel\Category">
-        <plugin name="category_move_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Category\Move"/>
-        <plugin name="category_delete_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Category\Remove"/>
+        <plugin name="group_plugin" type="Magento\CatalogUrlRewrite\Plugin\Model\Category\Store\Group"/>
     </type>
     <type name="Magento\Catalog\Block\Adminhtml\Product\Edit\Tab\Attributes">
         <plugin name="product_form_url_key_plugin" type="Magento\CatalogUrlRewrite\Plugin\Catalog\Block\Adminhtml\Product\Edit\Tab\Attributes"/>
