+++ app/code/Magento/CatalogUrlRewrite/etc/di.xml
@@ -17,12 +17,12 @@
         </arguments>
     </type>
     <type name="Magento\Catalog\Model\ResourceModel\Category">
-        <plugin name="category_move_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Category\Move"/>
-        <plugin name="category_delete_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Category\Remove"/>
-        <plugin name="update_url_path_for_different_stores" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Category\UpdateUrlPath"/>
+        <plugin name="category_move_plugin" type="Magento\CatalogUrlRewrite\Plugin\Model\Category\Move"/>
+        <plugin name="category_delete_plugin" type="Magento\CatalogUrlRewrite\Plugin\Model\Category\Remove"/>
+        <plugin name="update_url_path_for_different_stores" type="Magento\CatalogUrlRewrite\Plugin\Model\Category\UpdateUrlPath"/>
     </type>
     <type name="Magento\UrlRewrite\Model\StorageInterface">
-        <plugin name="storage_plugin" type="Magento\CatalogUrlRewrite\Model\Category\Plugin\Storage"/>
+        <plugin name="storage_plugin" type="Magento\CatalogUrlRewrite\Plugin\Model\Category\Storage"/>
     </type>
     <type name="Magento\CatalogUrlRewrite\Model\Map\UrlRewriteFinder">
         <arguments>
